+++ SensateIoT.Platform.Network.Common/Routing/CompositeRouter.cs
@@ -75,7 +75,7 @@ namespace SensateIoT.Platform.Network.Common.Routing
 				var result = Parallel.ForEach(messageList, this.InternalRoute);
 
 				if(!result.IsCompleted) {
-					this.m_logger.LogError("Unable to complete routing {count} messages.", messageList.Count);
+					this.m_logger.LogError("Unable to complete routing {count} messages", messageList.Count);
 				}
 			} catch(AggregateException ex) {
 				throw ex.InnerException!;
@@ -87,47 +87,104 @@ namespace SensateIoT.Platform.Network.Common.Routing
 		private void InternalRoute(IPlatformMessage message)
 		{
 			using(this.m_duration.NewTimer()) {
-				this.m_counter.Inc();
-				this.ProcessMessage(message);
+				var sensor = this.m_cache[message.SensorID];
+				this.ProcessMessage(sensor, message);
 			}
 		}
 
-		private void ProcessMessage(IPlatformMessage message)
+		private void ProcessMessage(Sensor sensor, IPlatformMessage message)
 		{
-			var sensor = this.m_cache[message.SensorID];
-
 			if(sensor == null) {
 				this.m_dropCounter.Inc();
-				this.m_logger.LogDebug("Unable to route message for sensor: {sensorId}. Sensor not found!", message.SensorID.ToString());
+				this.m_logger.LogDebug("Unable to route message for sensor: {sensorId}. Sensor not found", message.SensorID.ToString());
 				return;
 			}
 
-			this.m_logger.LogDebug("Routing message of type {type} for sensor {sensorId}.", message.Type.ToString("G"), sensor.ID.ToString());
+			this.m_logger.LogDebug("Routing message of type {type} for sensor {sensorId}", message.Type.ToString("G"), sensor.ID.ToString());
 			this.RouteMessage(message, sensor);
 		}
+
+		private bool IsValidSensor(Sensor sensor, Account account, ApiKey key)
+		{
+			bool invalid;
+
+			invalid = this.ValidateAccount(sensor, account, key);
+
+			invalid |= key.IsReadOnly;
+			invalid |= key.IsRevoked;
+
+			return !invalid;
+		}
+
+		private bool ValidateAccount(Sensor sensor, Account account, ApiKey key)
+		{
+			var invalid = false;
+
+			if(account.HasBillingLockout) {
+				this.m_logger.LogInformation("Skipping sensor {sensorId} due to billing lock", sensor.ID.ToString());
+				invalid = true;
+			}
+
+			if(account.IsBanned) {
+				this.m_logger.LogInformation("Skipping sensor because account {accountId:D} is banned", account.ID);
+				invalid = true;
+			}
+
+			invalid |= account.ID != key.AccountID;
+			return invalid;
+		}
+
 
 		private void RouteMessage(IPlatformMessage message, Sensor sensor)
 		{
 			var @event = CreateNetworkEvent(sensor);
+
+			this.m_counter.Inc();
 			message.PlatformTimestamp = DateTime.UtcNow;
 
+			if(!this.VerifySensor(sensor)) {
+				this.m_dropCounter.Inc();
+				this.m_eventQueue.EnqueueEvent(@event);
+				return;
+			}
+
 			foreach(var router in this.m_routers) {
-				bool @continue;
+				bool status;
 
 				try {
-					@continue = router.Route(sensor, message, @event);
+					status = router.Route(sensor, message, @event);
 				} catch(RouterException ex) {
-					@continue = false;
+					status = false;
 					this.m_dropCounter.Inc();
-					this.m_logger.LogWarning(ex, "Unable to route message for sensor {sensorId}", message.SensorID.ToString());
+					this.m_logger.LogWarning(ex, "Unable to route message for sensor {sensorId} using router {routerName}",
+											 message.SensorID.ToString(), router.Name);
 				}
 
-				if(!@continue) {
+				if(!status) {
+					this.m_logger.LogDebug("Routing cancelled by the {routerName}", router.Name);
 					break;
 				}
 			}
 
 			this.m_eventQueue.EnqueueEvent(@event);
+		}
+
+		private bool VerifySensor(Sensor sensor)
+		{
+			var account = this.m_cache.GetAccount(sensor.AccountID);
+			var key = this.m_cache.GetApiKey(sensor.SensorKey);
+
+			if(account == null) {
+				this.m_logger.LogWarning("Account with ID {accountId:D} not found", sensor.AccountID);
+				return false;
+			}
+
+			if(key == null) {
+				this.m_logger.LogWarning("API key for sensor with ID {sensorId} not found", sensor.ID);
+				return false;
+			}
+
+			return this.IsValidSensor(sensor, account, key);
 		}
 
 		private static NetworkEvent CreateNetworkEvent(Sensor sensor)